backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0
# redis>=4.5.2,<5.0.0  # Removed - using simple in-memory cache

# Vector Database (removed for direct tool provision)
# weaviate-client==3.25.3

# LangChain/LangGraph Stack
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langgraph>=0.1.0,<0.2.0
langsmith>=0.1.0,<0.2.0

# AI/ML & LLM Providers
openai>=1.30.0,<2.0.0
anthropic>=0.18.0,<1.0.0
httpx>=0.25.0,<1.0.0  # For DeepSeek API
numpy>=1.24.0,<2.0.0
# scikit-learn>=1.3.0,<2.0.0  # Only needed for vector operations
# sentence-transformers>=2.2.0,<3.0.0  # Only needed for embeddings

# Vector Stores (removed for direct tool provision)
# chromadb>=0.4.0,<0.5.0
# faiss-cpu>=1.7.0,<2.0.0

# Background Tasks - Celery removed (Redis dependency)
# celery==5.3.4  # Removed - using simple background tasks
# celery[redis]==5.3.4  # Removed - using simple background tasks

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
# opentelemetry-instrumentation-redis==0.42b0  # Removed - no Redis
prometheus-client==0.19.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Browser Automation
playwright==1.40.0

# Model Context Protocol
mcp>=1.0.0

# Utilities
python-jose[cryptography]==3.3.0
cryptography>=41.0.0,<42.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
typer==0.9.0
rich==13.7.0

# GraphQL
strawberry-graphql[fastapi]==0.214.0

# Validation & Serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Scheduling
apscheduler==3.10.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0